requirements.txt
asgiref==3.2.7
Django==3.0.4
django-extensions==2.2.9
django-filter==2.2.0
django-rest-passwordreset==1.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Markdown==3.2.1
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
